SampleTable.cpp: Fix a regression caused by a fix for bug 28076789.
Detail: Before the original fix (Id207f369ab7b27787d83f5d8fc48dc53ed9fcdc9)
for 28076789, the code allowed a time-to-sample table size to be 0.
The change made in that fix disallowed such situation, which in fact should be allowed.
This current patch allows it again while maintaining the security of the previous fix.

Bug: 28288202
Bug: 28076789
Change-Id: I1c9a60c7f0cfcbd3d908f24998dde15d5136a295